# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Black Pear Joggers API",
        "contact": {
            "name": "Paul Evans",
            "url": "https://bpj.org.uk",
            "email": "paul3vanz@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/performances": {
            "get": {
                "tags": [
                    "Performances"
                ],
                "summary": "Get all performances",
                "operationId": "a86695eb918939d2371e2cd180904ce8",
                "parameters": [
                    {
                        "name": "athleteId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/performances/{id}": {
            "get": {
                "tags": [
                    "Performances"
                ],
                "summary": "Get a performance by ID",
                "operationId": "b28c5417f233f62e3b7c68e1d6787025",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rankings/{athleteId}": {
            "get": {
                "tags": [
                    "Rankings"
                ],
                "summary": "Get an athlete's runBritain rankings history",
                "operationId": "21b7166e3d4a036ed0f9db2b89658e28",
                "parameters": [
                    {
                        "name": "athleteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rankings/{athleteId}/{year}": {
            "get": {
                "tags": [
                    "Rankings"
                ],
                "summary": "Get an athlete's runBritain rankings history for a year",
                "operationId": "1f099996e70d97ef0520307b87f9593c",
                "parameters": [
                    {
                        "name": "athleteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/records": {
            "get": {
                "tags": [
                    "Records"
                ],
                "summary": "Get club records",
                "operationId": "42712d55c7ea8e69618967e62e7d7c4d",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "M",
                                "W"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "event",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "distance",
                                "gender"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/records/query": {
            "post": {
                "tags": [
                    "Records"
                ],
                "summary": "Flag a club record performance for review",
                "operationId": "277d81c930a5553a184fa7b2666c19c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "record",
                                    "reason"
                                ],
                                "properties": {
                                    "record": {
                                        "type": "object"
                                    },
                                    "reason": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}